Hackish implementation for listening to changes on Nepomuk Resources.
At the moment, this is implemented by a signleton that listens to all statementAdded() and statementDeleted() calls on Nepomuk, which
can then call back to registered objects if the statement effects an interesting resource.

NB This is only a temporary measure until the Nepomuk folks implement listening for changes better on their side.

svn path=/trunk/playground/network/telepathy-contactlist/; revision=1100114