Make Telepathy-Contactlist pick up changes in realtime.
This patch uses the awesome QueryServiceClient to monitor sparql queries for further results.

This means telepathy-contactlist is almost usable right now: added contacts are shown instantly
as they get approved, and removed contacts are thrown away as they are removed as well.

For finishing, now tp-contactlist only shows contacts that published their presence to us,
in the same flavour empathy does.
In the end, this is probably not really what we'll want to do, but it's here to show it works.

With the current architecture, it's just a matter of tuning the SPARQL query retrieving contacts for
deciding what will be shown: all the rest is done by magic.

Signed-off-by: Dario Freddi <dario.freddi@collabora.co.uk>

svn path=/trunk/playground/network/telepathy-contactlist/; revision=1116075