Maintain a list of recently used rooms in JoinChatroomDialog
Upon joining a room (more exact upon clicking "Ok") the room is saved in
the recently used list for this account.
ATM it remembers up to 8 rooms and single rooms can be removed as well
as the whole list can be cleared.

The complete list is shown in a new tab and the rooms are used for
completion in the lineEdit.

CCBUG: 291711
REVIEW: 104269