Set the new presence first and only then remove the temp presence
Removing the temp presence first causes setting new selected index (as the currently selected one was just deleted) which is set to the next in the list, which in case of the temp presence is the 'Now playing...' one, which prompts the user for enabling it if it is disabled.
So now we first change the presence (and current index) and only then remove the temp presence, then it won't cause another unwanted index re-setting.

Reviewed-by: David Edmundson

BUG: 304085
FIXED-IN: 0.5