Don't remove and insert a temp presence, but only update it if it exists
Inserting the temp presence first and then removing it afterwards causes indexes shift, which in turn can cause problems, like accidental "now playing..." presence being selected and activated and then being deactivated right away when the old temp presence was removed and indexes were shifted again.
This is a clearer approach, which only updates the temp presence if it is already being used.

Reviewed-by: David Edmundson