Handle the uninitialized presence chooser state differently
This moves part of the logic to KTp::GlobalPresence which now
sets Unknown presence before it's properly initialized. We catch
this here and show an empty combobox with a spinner. This fixes
the case where the presence was Offline, but the presence chooser
was empty and spinning away happily. After GlobalPresence is
initialized with AccountManager, it emits presenceChanged with
Offline presence, changing the combobox value properly.

Reviewed-by: David Edmundson