Work around bug that fails to display local-xmpp contacts.
Work around https://bugs.freedesktop.org/show_bug.cgi?id=57739.

Due to a bug in TelepathyQt if we request a feature that the connection does not support, it simply fails.

Local-XMPP does not support grouping, so salut the local-xmpp spec does not either.
When we request the ContactRosterGroups feature on all connections, salut fails and therefore doesn't show any contacts.

Fetching ContactRosterGroups seperately is not a viable option as contacts would move after loading

In this hack we make a new ConnectionFactory that fetches ContactRosterGroups on all connections _except_ salut

REVIEW: 110265