Fix for a bug that caused the contacts list to not expand some groups on opening and filtering
I've fixed the 325506 bug by creating a decision upon the validity of the inserted item's parent. If it is valid, it works as before, considering it is a top-level element we want to collapse/expand. If it is invalid, the inserted item itself is probably the top-level node (hence the invalid parent - there's none), so we get it in another way.

I don't think the second validity check is strictly needed, but I thought it was safer to keep it anyway.

REVIEW: 117791
BUG: 325506