Use onRequestedPresenceChanged in the contact list
Use requestedPresenceChanged in the presence chooser, as well as calling onPresenceChanged after every selectable option that can trigger config changes or presence changes in onUserActivatedComboChange. This improves the contact list interaction with #116940. The selected presence is instantly activated instead of the last selected user presence when deactivating the nowPlaying status message plugin.

REVIEW: 117366